use uint32 for nonce and uint64 for timeout timestamp

+++ solidity/script/7683/SubmitOrder.s.sol
@@ -41,7 +41,7 @@ contract DeployScript is Script {
                     amount, // amountOut
                     localDomain, // sourceDomain
                     destinationDomain, // destinationDomain
-                    block.timestamp + 1 days, // timeoutTimestamp
+                    uint64(block.timestamp + 1 days), // timeoutTimestamp
                     data
                 )
             )
